code refactor - moved header to common folder in components
